Import chromium-browser_48.0.2564.82.orig.tar.xz from Ubuntu
All .gitignore files dropped, as this would have prevented much of the
tarball being imported.
 # find . -name .gitignore -delete

Removed files over 50mb in size to make github happy.
 # find . -type f -size +49M -delete

+++ cloud_print/service/win/service_controller.cc
@@ -8,6 +8,7 @@
 #include <atlcom.h>
 #include <atlctl.h>
 
+#include "base/base_switches.h"
 #include "base/command_line.h"
 #include "base/files/file_path.h"
 #include "base/files/file_util.h"
@@ -20,6 +21,7 @@
 #include "cloud_print/service/win/chrome_launcher.h"
 #include "cloud_print/service/win/local_security_policy.h"
 #include "cloud_print/service/win/service_utils.h"
+#include "content/public/common/content_switches.h"
 
 namespace {
 
@@ -298,7 +300,8 @@ void ServiceController::UpdateState() {
   std::vector<uint8> buffer(config_size, 0);
   QUERY_SERVICE_CONFIG* config =
       reinterpret_cast<QUERY_SERVICE_CONFIG*>(&buffer[0]);
-  if (!::QueryServiceConfig(service.Get(), config, buffer.size(),
+  if (!::QueryServiceConfig(service.Get(), config,
+                            static_cast<DWORD>(buffer.size()),
                             &config_size) ||
       config_size != buffer.size()) {
     return;
